services/bot/requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
python-dotenv==1.0.0